all user functionality working, ready for tickets

+++ controller/userController.js
@@ -3,32 +3,31 @@ const router = express.Router();
 const userService = require('../service/userService');
 const logger = require("../util/logger");
 const jwt = require("jsonwebtoken");
-//const bcrypt = require("bcrypt");
-
-
+const { authenticateToken } = require("../util/jwt");
 require('dotenv').config();
 const SECRET_KEY = process.env.SECRET_KEY;
-
-//router.use(express.json());
 
 
 router.post("/login", async (req, res) => {
     const {username, password} = req.body;
+    console.log(username, password);
+    if(!username || !password){
+        return res.status(400).json({message: "bad request, please try again"})
+    }
     const data = await userService.validateLogin(username, password);
-    
-    
+      
     if(data){
         // req.session.username = username;
+        //console.log(data);
         
-        user = data.user;
-        console.log(user.is_manager);
-        console.log(user.UserId);
-        console.log(user.username);
+        //console.log(user.is_manager);
+        //console.log(user.UserId);
+        //console.log(user.username);
         const token = jwt.sign(
             {
-                id: user.UserId,
+                id: data.UserId,
                 username: username,
-                is_manager: user.is_manager
+                is_manager: data.is_manager
             },
                 SECRET_KEY,
             {
@@ -51,10 +50,10 @@ router.get("/", (req, res) => {
 router.get("/:userId", async (req, res) => {
     const id = req.params.userId;
     console.log(id);
-    //console.log(userService.getUser(id));
+    
     const user = await userService.getUser(id);
     res.status(200).json(user);
-    //res.status(200).send("This is the root users route");
+    
     logger.info(`Get request made: id = ${id}`);
 });
 
@@ -62,10 +61,13 @@ router.get("/:userId", async (req, res) => {
 router.get("/username/:username", async (req, res) => {
     const username = req.params.username;
     console.log(username);
-    //console.log(userService.getUser(id));
+    
     const response = await userService.getUserByUsername(username);
+    if(!response){
+        return res.status(400).json({"message": "Invalid request"})
+    }
     res.status(200).json(response);
-    //res.status(200).send("This is the root users route");
+    
     logger.info(`Get request made: username = ${username}`);
 });
 
@@ -82,50 +84,48 @@ router.post("/", async (req, res) => {
         res.status(400).json({Message: 
             "User not created, please ensure your username is unique and that you have a valid password"});
     }
-    
-    
-    //res.status(201).json({message: "Item Created!", item: jsonData});
+      
 });
 
-router.put("/", async (req, res) => {
+router.put("/", authenticateToken, async (req, res) => {
+    console.log(req.user.is_manager,req.user.username,req.body.username);
+    if(!req.user.is_manager && !(req.user.username == req.body.username)){
+        return res.status(403).json({"message": "Invalid permissions to update this account."})
+    }
     const jsonData = req.body;
+    
     console.log(jsonData);
     const msg = await userService.updateUser(jsonData);
     res.status(200).json(msg);
 });
 
 
-router.put("/managerstatus", async (req, res) =>{
+router.put("/managerstatus",authenticateToken, async (req, res) =>{
+    console.log(req.user.is_manager);
+    if(!req.user.is_manager){
+        return res.status(403).json({message: "Forbidden access"});
+    }
     const jsonData = req.body;
     console.log(jsonData);
     const msg = await userService.updateManagerStatus(jsonData.UserId, jsonData.is_manager);
+    if(!msg){
+        return res.status(400).json({"message": "Bad request, please include is_manager(boolean)"})
+    }
     res.status(200).json(msg);
 });
 
 
-router.delete("/:userId", async (req, res) => {
+router.delete("/:userId", authenticateToken, async (req, res) => {
+    if(!req.user.is_manager){
+        return res.status(403).json({message: "You do not have access to this route!"});
+    }
     const id = req.params.userId;
     console.log(id);
-    //console.log(userService.getUser(id));
+    
     const msg = await userService.deleteUser(id);
     res.status(200).json(msg);
-    //res.status(200).send("This is the root users route");
     logger.info(`Delete request made: id = ${id}`);
 });
-
-  function validateItemMiddleware(req, res, next){
-    // check if there is a valid name and price
-    const jsonBody = req.body;
-
-    if(validateItem(jsonBody)){
-        next();
-    }else{
-        res.status(400).json({
-            message: "Invalid Name or Price"
-        });
-    }
-};
-
 
 
 module.exports = router;
